Refactor a bit the pkg/realtime package
